move model to new lib layout
