+++ projects/storefrontlib/src/cms-structure/page/component/component-wrapper.directive.spec.ts
@@ -7,7 +7,7 @@ import {
 } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { ComponentWrapperDirective } from './component-wrapper.directive';
-import { CmsComponentData } from '../../../lib/cms/components/cms-component-data';
+import { CmsComponentData } from '../model/cms-component-data';
 import {
   CmsComponent,
   CmsService,
