+++ projects/storefrontlib/src/cms-structure/page/component/component-wrapper.directive.ts
@@ -18,7 +18,7 @@ import {
   ComponentMapperService,
   CxApiService
 } from '@spartacus/core';
-import { CmsComponentData } from '../../../lib/cms/components/cms-component-data';
+import { CmsComponentData } from '../model/cms-component-data';
 import { isPlatformServer } from '@angular/common';
 
 @Directive({
