+++ projects/storefrontlib/src/cms-structure/page/index.ts
@@ -1,3 +1,6 @@
+// model
+export * from './model/index';
+
 // slot
 export * from './slot/page-slot.module';
 export * from './slot/page-slot.component';
