+++ projects/storefrontlib/src/cms-structure/page/model/cms-component-data.ts
