+++ projects/storefrontlib/src/cms-structure/page/model/index.ts
@@ -1,0 +1,1 @@
+export * from './cms-component-data';
