+++ projects/storefrontlib/src/lib/cms-lib/category-navigation/category-navigation.component.spec.ts
@@ -7,7 +7,7 @@ import createSpy = jasmine.createSpy;
 import { Component as SpaComponent } from '@spartacus/core';
 import { NavigationComponentService } from '../navigation/navigation.component.service';
 import { CategoryNavigationComponent } from './category-navigation.component';
-import { CmsComponentData } from '../../cms/components/cms-component-data';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 import { NavigationNode } from '../navigation/navigation-node.model';
 
 @Component({
