+++ projects/storefrontlib/src/lib/cms-lib/mini-cart/mini-cart.component.ts
@@ -4,7 +4,7 @@ import { Observable } from 'rxjs';
 
 import { CartService, CmsMiniCartComponent, Cart } from '@spartacus/core';
 
-import { CmsComponentData } from './../../cms/components/cms-component-data';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 
 @Component({
   selector: 'cx-mini-cart',
