+++ projects/storefrontlib/src/lib/cms-lib/navigation/navigation.module.ts
@@ -12,7 +12,7 @@ import {
 import { NavigationComponent } from './navigation.component';
 import { NavigationUIComponent } from './navigation-ui.component';
 import { NavigationComponentService } from './navigation.component.service';
-import { CmsComponentData } from '../../cms/components/cms-component-data';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 
 @NgModule({
   imports: [
