+++ projects/storefrontlib/src/lib/cms-lib/paragraph/paragraph.component.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { CmsComponentData } from '../../cms/components/cms-component-data';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 import { CmsParagraphComponent } from '@spartacus/core';
 
 @Component({
