+++ projects/storefrontlib/src/lib/cms-lib/site-context-selector/site-context-component.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, Optional, Injector } from '@angular/core';
-import { CmsComponentData } from '../../cms/components/cms-component-data';
+import { CmsComponentData } from '../../../cms-structure/page/model/cms-component-data';
 import {
   CmsSiteContextSelectorComponent,
   ContextServiceMap,
