+++ projects/storefrontlib/src/lib/cms/index.ts
@@ -1,5 +1,4 @@
 export * from './cms.module';
-export * from './components/cms-component-data';
 export * from './guards/cms-page.guard';
 export * from './page-layout/page-layout.module';
 export * from './page-layout/page-layout.component';
